ceph-volume-ansible-prs: deploy the target branch, not the source branch
We're going to start deploying the target branch, the branch we're
merging PRs into, instead of the source branch. This way we don't have to
wait on packages as we can assume there is somethign already available
for the target branch. The ceph-volume code we're testing will be
rsynced to the testing nodes from the source branch.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>